add "clickToOpen" property with default true
This will disable the toggle action when a panel is clicked leaving only the "open" property to determine whether a panel is open or not. This was for a use case where the panel being open or not was directly tied to the current route. I was using a custom action in the panel toggle component to handle switching routes.